# public/swagger.yaml
swagger: "2.0"
info:
  title: 게시판
paths:
...
  /api/comment/deleteComment:
    post:
      summary: (댓글 삭제)
      tags:
        - comment
      parameters:
        - name: body
          in: body
          description: "삭제할 댓글 정보"
          required: true
          schema:
            $ref: '#/comment/deleteComment/request'

      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/comment/deleteComment/response_200'
        '400':
          description: "에러"
          schema:
            $ref: '#/comment/response_400'
  /api/comment/modifyComment:
    post:
      summary: (댓글 수정)
      tags:
        - comment

      parameters:
        - name: body
          in: body
          description: "수정할 데이터"
          required: true
          schema:
            $ref: '#/comment/modifyComment/request'
      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/comment/modifyComment/response_200'
        '400':
          description: "에러"
          schema:
            $ref: '#/comment/response_400'
  /api/comment/saveRootComment:
    post:
      summary: (부모댓글 저장)
      tags:
        - comment

      parameters:
        - name: body
          in: body
          description: "저장할 댓글 데이터"
          required: true
          schema:
            $ref: '#/comment/saveRootComment/request'
      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/comment/saveRootComment/response_200'
        '400':
          description: "에러"
          schema:
            $ref: '#/comment/response_400'
  /api/comment/saveChildComment:
    post:
      summary: (자식댓글 저장)
      tags:
        - comment

      parameters:
        - name: body
          in: body
          description: "저장할 댓글 데이터"
          required: true
          schema:
            $ref: '#/comment/saveChildComment/request'
      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/comment/saveChildComment/response_200'
        '400':
          description: "에러"
          schema:
            $ref: '#/comment/response_400'
  /api/comment/getCommentPage:
      post:
        summary: (등록순으로 댓글 요청)
        tags:
          - comment

        parameters:
          - name: body
            in: body
            description: "요청할 댓글 정보"
            required: true
            schema:
              $ref: '#/comment/getCommentPage/request'
        responses:
          '200':
            description: "성공"
            schema:
              $ref: '#/comment/getCommentPage/response_200'
          '400':
            description: "에러"
            schema:
              $ref: '#/comment/response_400'
  /api/comment/getLatestComment:
        post:
          summary: (최신순으로 댓글 요청)
          tags:
            - comment

          parameters:
            - name: body
              in: body
              description: "요청할 댓글 정보"
              required: true
              schema:
                $ref: '#/comment/getLatestComment/request'
          responses:
            '200':
              description: "성공"
              schema:
                $ref: '#/comment/getLatestComment/response_200'
            '400':
              description: "에러"
              schema:
                $ref: '#/comment/response_400'



####### 신고 #######
  /api/report/saveReport:
    post:
      summary: (게시글 신고)
      tags:
        - report
      parameters:
        - name: body
          in: body
          description: "신고할 게시글 정보"
          required: true
          schema:
            $ref: '#/report/saveReport/request'

      responses:
        '이미 신고한 게시물 200':
          description: "성공"
          schema:
            $ref: '#/report/saveReport/response_200_case_1'

        '신고 완료 200':
          description: "성공"
          schema:
            $ref: '#/report/saveReport/response_200_case_2'
        
        '400':
          description: "에러"
          schema:
            $ref: '#/report/response_400'
  /api/report/getReportPost:
    post:
      summary: (신고된 게시글 요청)
      tags:
        - report
      parameters:
        - name: body
          in: body
          description: "신고된 게시글 정보"
          required: true
          schema:
            $ref: '#/report/getReportPost/request'

      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/report/getReportPost/response_200'

        '400':
          description: "에러"
          schema:
            $ref: '#/report/response_400'
  /api/report/deleteReportPost:
    post:
      summary: (신고된 게시글 삭제)
      tags:
        - report
      parameters:
        - name: body
          in: body
          description: "삭제할 게시글 정보"
          required: true
          schema:
            $ref: '#/report/deleteReportPost/request'

      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/report/deleteReportPost/response_200'

        '400':
          description: "에러"
          schema:
            $ref: '#/report/response_400'
  /api/report/saveReportComment:
    post:
      summary: (댓글 신고)
      tags:
        - report
      parameters:
        - name: body
          in: body
          description: "신고할 댓글 정보"
          required: true
          schema:
            $ref: '#/report/saveReportComment/request'

      responses:
        '이미 신고한 게시물 200':
          description: "성공"
          schema:
            $ref: '#/report/saveReportComment/response_200_case_1'
        
        '신고 완료 200+':
          description: "성공"
          schema:
            $ref: '#/report/saveReportComment/response_200_case_2'

        '400':
          description: "에러"
          schema:
            $ref: '#/report/response_400'
  /api/report/getReportComment:
    post:
      summary: (신고된 댓글 요청)
      tags:
        - report
      parameters:
        - name: body
          in: body
          description: "신고된 댓글 정보"
          required: true
          schema:
            $ref: '#/report/getReportComment/request'

      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/report/getReportComment/response_200'

        '400':
          description: "에러"
          schema:
            $ref: '#/report/response_400'
  /api/report/deleteReportComment:
    post:
      summary: (신고된 댓글 삭제)
      tags:
        - report
      parameters:
        - name: body
          in: body
          description: "삭제할 댓글 정보"
          required: true
          schema:
            $ref: '#/report/deleteReportComment/request'

      responses:
        '200':
          description: "성공"
          schema:
            $ref: '#/report/deleteReportComment/response_200'

        '400':
          description: "에러"
          schema:
            $ref: '#/report/response_400'


####### 보드 #######
...